+++ jest.config.js
@@ -1,6 +1,6 @@
 module.exports = {
   transform: {'^.+\\.ts?$': 'ts-jest'},
   testEnvironment: 'node',
-  testRegex: '/test/.*\\.(test|spec)?\\.js$',
+  testRegex: '/test/.*\\.(test|spec)?\\.ts$',
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
 };