+++ src/Constants.ts
@@ -1,4 +1,0 @@
-export abstract class Constants {
-  static readonly INVALID_DID: string = 'invalidDid';
-  static readonly URL_PATHNAME_REGEX = /.+?:\/\/.+?(\/.+?)(?:#|\?|$)/;
-}