+++ src/config.ts
@@ -1,7 +1,0 @@
-/** Registrar URL's*/
-export const registrarUrlCreate = 'https://uniregistrar.io/1.0/create';
-export const registrarUrlUpdate = 'https://uniregistrar.io/1.0/update';
-export const registrarUrlDeactivate = 'https://uniregistrar.io/1.0/deactivate';
-
-/** Resolver URL's*/
-export const resolverUrlResolve = 'https://dev.uniresolver.io/1.0/identifiers';