+++ src/index.ts
@@ -1,3 +1,4 @@
-export { Resolver } from './resolver/Resolver';
-export { Registrar } from './registrar/Registrar';
-export { CrudRequestBuilder } from './registrar/rest/CrudRequestBuilder';
+export { UniResolver } from './resolver/UniResolver';
+export { UniRegistrar } from './registrar/UniRegistrar';
+export { DIDRegistrationRequestBuilder } from './registrar/DIDRegistrationRequestBuilder';
+export { DIDRegistrationResult } from './types/types';