+++ src/registrar/DIDRegistrationRequest.ts
@@ -2,15 +2,15 @@
 
 import { DIDDocument } from 'did-resolver';
 
-import { ICrudRequest } from './ICrudRequest';
+import { IDIDRegistrationRequest } from '../types/types';
 
-export class CrudRequest implements ICrudRequest {
+export class DIDRegistrationRequest implements IDIDRegistrationRequest {
   readonly jobId: string;
   readonly didDocument: DIDDocument;
   readonly options: Record<string, unknown>;
   readonly secret: Record<string, unknown> | string;
 
-  protected constructor(request?: CrudRequest) {
+  protected constructor(request?: DIDRegistrationRequest) {
     if (request) {
       Object.assign(this, request);
     }