+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Release Notes
 
+## v0.3.0 - 2021-10-09
+WARNING: API changed. Many methods and classnames have changed!
+
+Improved API, results have types. All main methods return a promise and are configured using a configuration object
+
+- Added:
+    * Introduction of a configuration object
+    * Allow registration of multiple DID methods when used as DIF resolver
+
+- Changed:
+    * All registration/resolution methods are typed and return a promise
+    * No more getURL methods in the resolver/registrar. Use getConfig() to access the values
+    * Prefix the Resolver and Registrar classes with Uni, to not have to alias them on imports with DIF resolver present
+
 
 ## v0.2.0 - 2021-09-13
 Improved error handling + DIF resolver support, which means you can now use this the resolver as a DIF did-client supported resolver.